Implementation of Final Level
In this commit the final or boss level has been implemented. This level consists of beating the Final Boss by answering multiple choice questions correctly. If you answer more than half of the questions correctly, you will win. However, if it is less, you loose. The concept of the final level is completely different to the other levels as well as the graphics.

To achieve this, several methods, mirroring  the standard ones from the other levels have been added,  some just differing subtly (main/final_level(), interact_logic/finalInteract() and interact_logic/drawCustomText()) and not so subtle such as (display_logic/refreshFinal() and interact_logic/next_question()). On the graphical aspect, a new folder containing all the new models for this spcecial level have been added in Assests/boss.